% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movenetapp-package.R
\docType{package}
\name{movenetapp-package}
\alias{movenetapp}
\alias{movenetapp-package}
\title{movenetapp: Wrapper For Movenet Shiny App}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Carlijn Bogaardt \email{Carlijn.Bogaardt@glasgow.ac.uk}

Other contributors:
\itemize{
  \item Matt Denwood \email{md@sund.ku.dk} [contributor]
}

}
\keyword{internal}
